Extract AbstractParquetDataSource
